pyproject.toml
[project]
name = "openpyxl-with-excel"
version = "0.1.0"
description = ""
authors = [
    {name = "Tumininu",email = "anisegit@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "openpyxl_with_excel", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
